checking and bug fix and optimize after live server upload

+++ database/migrations/2024_09_14_102935_create_user_statuses_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->text('reason');
             $table->integer('blocked_duration')->nullable(); // in hours
             $table->timestamp('blocked_resolved_request_at')->nullable();
-            $table->float('blocked_resolved_charge')->nullable();
+            $table->decimal('blocked_resolved_charge', 10, 2)->nullable();
             $table->timestamp('resolved_at')->nullable();
             $table->integer('created_by')->nullable();
             $table->integer('updated_by')->nullable();
